+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Tracks qoqo-qryd changes after 0.5
 
 # 0.11.5
 
-* Modified `TweezerDevice.from_api()` endpoint
+* Modified `TweezerDevice.from_api()` endpoint, default device name for better `APIBackend` support
 * Added support for `PragmaActiveReset` for `APIBackend`, `TweezerDevice`
 
 # 0.11.4